refactor(cl): Yank new bte code to simplify first PR
