+++ cl/recap/tasks.py
@@ -60,8 +60,7 @@ from cl.recap.models import (
     REQUEST_TYPE,
     UPLOAD_TYPE,
 )
-from cl.scrapers.tasks import extract_recap_pdf
-from cl.scrapers.transformer_extractor_utils import get_page_count
+from cl.scrapers.tasks import extract_recap_pdf, get_page_count
 from cl.search.models import Docket, DocketEntry, RECAPDocument
 from cl.search.tasks import add_or_update_recap_docket, add_items_to_solr
 
@@ -305,7 +304,8 @@ def process_recap_pdf(self, pk):
             rd.filepath_local.save(file_name, cf, save=False)
 
             # Do page count and extraction
-            rd.page_count = get_page_count(rd.filepath_local.path)["pg_count"]
+            extension = rd.filepath_local.path.split(".")[-1]
+            rd.page_count = get_page_count(rd.filepath_local.path, extension)
             rd.file_size = rd.filepath_local.size
 
         rd.ocr_status = None
