+++ cl/scrapers/tasks.py
@@ -13,6 +13,11 @@ from django.apps import apps
 from django.conf import settings
 from django.core.files.base import ContentFile
 from django.db import IntegrityError
+from django.utils.encoding import (
+    smart_text,
+    DjangoUnicodeDecodeError,
+    force_text,
+)
 from django.utils.timezone import now
 from lxml.etree import XMLSyntaxError
 from lxml.html.clean import Cleaner
@@ -40,6 +45,8 @@ from cl.scrapers.transformer_extractor_utils import (
 from cl.search.models import Opinion, RECAPDocument, Docket
 from cl.search.tasks import add_items_to_solr
 from juriscraper.pacer import PacerSession, CaseQuery
+
+DEVNULL = open("/dev/null", "w")
 
 logger = logging.getLogger(__name__)
 
@@ -402,7 +409,6 @@ def extract_recap_pdf(pks, skip_ocr=False, check_if_needed=True):
 
 
 @app.task
-
 def rasterize_pdf(path, destination):
     """Convert the PDF into a multipage Tiff file.
 
@@ -475,91 +481,6 @@ def extract_by_ocr(path: str) -> (bool, str):
         txt = cleanup_ocr_text(txt)
 
     return True, txt
-
-
-def set_mp3_meta_data(audio_obj, mp3_path):
-    """Sets the meta data on the mp3 file to good values.
-
-    :param audio_obj: an Audio object to clean up.
-    :param mp3_path: the path to the mp3 to be converted.
-    """
-    court = audio_obj.docket.court
-
-    # Load the file, delete the old tags and create a new one.
-    audio_file = eyed3.load(mp3_path)
-
-    # Undocumented API from eyed3.plugins.classic.ClassicPlugin#handleRemoves
-    id3.Tag.remove(
-        audio_file.tag.file_info.name,
-        id3.ID3_ANY_VERSION,
-        preserve_file_time=False,
-    )
-    audio_file.initTag()
-    audio_file.tag.title = best_case_name(audio_obj)
-    audio_file.tag.album = "{court}, {year}".format(
-        court=court.full_name, year=audio_obj.docket.date_argued.year
-    )
-    audio_file.tag.artist = court.full_name
-    audio_file.tag.artist_url = court.url
-    audio_file.tag.audio_source_url = audio_obj.download_url
-    audio_file.tag.comments.set(
-        "Argued: {date_argued}. Docket number: {docket_number}".format(
-            date_argued=audio_obj.docket.date_argued.strftime("%Y-%m-%d"),
-            docket_number=audio_obj.docket.docket_number,
-        )
-    )
-    audio_file.tag.genre = "Speech"
-    audio_file.tag.publisher = "Free Law Project"
-    audio_file.tag.publisher_url = "https://free.law"
-    audio_file.tag.recording_date = audio_obj.docket.date_argued.strftime(
-        "%Y-%m-%d"
-    )
-
-    # Add images to the mp3. If it has a seal, use that for the Front Cover
-    # and use the FLP logo for the Publisher Logo. If it lacks a seal, use the
-    # Publisher logo for both the front cover and the Publisher logo.
-    try:
-        has_seal = seals_data[court.pk]["has_seal"]
-    except AttributeError:
-        # Unknown court in Seal Rookery.
-        has_seal = False
-    except KeyError:
-        # Unknown court altogether (perhaps a test?)
-        has_seal = False
-
-    flp_image_frames = [
-        3,  # "Front Cover". Complete list at eyed3/id3/frames.py
-        14,  # "Publisher logo".
-    ]
-    if has_seal:
-        with open(
-            os.path.join(seals_root, "512", "%s.png" % court.pk), "rb"
-        ) as f:
-            audio_file.tag.images.set(
-                3, f.read(), "image/png", "Seal for %s" % court.short_name
-            )
-        flp_image_frames.remove(3)
-
-    for frame in flp_image_frames:
-        with open(
-            os.path.join(
-                settings.INSTALL_ROOT,
-                "cl",
-                "audio",
-                "static",
-                "png",
-                "producer-300x300.png",
-            ),
-            "rb",
-        ) as f:
-            audio_file.tag.images.set(
-                frame,
-                f.read(),
-                "image/png",
-                "Created for the public domain by Free Law Project",
-            )
-
-    audio_file.tag.save()
 
 
 @app.task
